--- crates/common/mapper_utils/Cargo.toml
@@ -1,36 +1,0 @@
-[package]
-name = "mapper_utils"
-description = "thin-edge aws extension adding support for aws data model"
-version = { workspace = true }
-authors = { workspace = true }
-edition = { workspace = true }
-rust-version = { workspace = true }
-license = { workspace = true }
-homepage = { workspace = true }
-repository = { workspace = true }
-
-[dependencies]
-async-trait = "0.1"
-clock = { path = "../../common/clock" }
-log = "0.4"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-tedge_actors = { path = "../../core/tedge_actors" }
-tedge_api = { path = "../../core/tedge_api" }
-tedge_config = { path = "../../common/tedge_config" }
-tedge_mqtt_ext = { path = "../../extensions/tedge_mqtt_ext" }
-tedge_utils = { path = "../../common/tedge_utils" }
-thiserror = "1.0"
-time = "0.3"
-tokio = { version = "1.23", features = ["macros"] }
-toml = "0.5"
-tracing = { version = "0.1", features = ["attributes", "log"] }
-
-[dev-dependencies]
-assert-json-diff = "2.0"
-assert_matches = "1.5"
-rand = "0.8"
-serde = "1.0"
-serde_json = "1.0"
-time = { version = "0.3", features = ["macros"] }
-tokio = { version = "1.23", features = ["time"] }
