+++ crates/extensions/az_mapper_ext/Cargo.toml
@@ -13,7 +13,6 @@ repository = { workspace = true }
 async-trait = "0.1"
 clock = { path = "../../common/clock" }
 log = "0.4"
-mapper_utils = { path = "../../common/mapper_utils" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 tedge_actors = { path = "../../core/tedge_actors" }
