+++ crates/extensions/az_mapper_ext/src/lib.rs
@@ -1,1 +1,3 @@
 pub mod converter;
+pub mod error;
+pub mod size_threshold;
